backend/requirements.txt
anyio==3.6.2
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
autopep8==2.0.0
backports.zoneinfo==0.2.1
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
click==8.1.3
constantly==15.1.0
cryptography==39.0.1
daphne==4.0.0
Django==4.1.2
django-cors-headers==3.13.0
django-filter==22.1
django-mysql==4.8.0
django-picklefield==3.1
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
h11==0.14.0
httptools==0.5.0
hyperlink==21.0.0
idna==3.4
importlib-metadata==5.0.0
incremental==22.10.0
Markdown==3.4.1
msgpack==1.0.4
Pillow==9.3.0
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==23.0.0
python-dotenv==1.0.0
pytz==2022.6
PyYAML==6.0
redis==4.5.1
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
tomli==2.0.1
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
tzdata==2022.7
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.2
zipp==3.10.0
zope.interface==5.5.2